requirements.txt
langchain
langchain_openai
langchain_core
langchain_community
streamlit
sse_starlette
python-dotenv
langserve
fastapi
uvicorn
bs4
ipykernel
pypdf
chromadb
faiss-cpu
groq
cassio
langchain-groq
wikipedia
arxiv
langchainhub